backend/core/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.3
gunicorn==21.2.0
dj-database-url==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
whitenoise==6.6.0